Serialized geofence flows and hardened the Yale token and lock fallback paths
